requirements.txt
Flask==3.1.1
PyPDF2==3.0.1
pdf2docx==0.5.8
python-pptx==1.0.2
reportlab==4.4.3
Werkzeug==3.1.3
gunicorn==21.2.0
bcrypt
flask-login
flask-migrate
pandas
pdfplumber
python-docx
qrcode[pil]
docx2pdf
flask-sqlalchemy